Simplify `BaseCoordinateFrame._frameattr_equiv()`
If both inputs are `BaseRepresentationOrDifferential` instances then the
code now makes use of their `__eq__` method instead of comparing all
their components itself. The code now takes a slightly different path if
`type(right_fattr)` is a strict subclass of `type(left_fattr)` to
prevent `BaseRepresentationOrDifferential.__eq__()` from raising a
`TypeError`, but that should not be a problem in practice because the
code still falls back to comparing the Cartesian representations of the
inputs if the classes do not match exactly. Furthermore, the code path
taken no longer depends on the order of the function inputs.