Address Ruff rule PLR0911 in `coordinates`
The rule guards against functions with very many `return` statements.
There is one function in `coordinates` that violates the rule. It has
been updated to have fewer `return` statements, but it still has too
many. It is possible to make it comply, but it is not obvious if that
would benefit readability. The rule is therefore addressed by adding a
`noqa` instruction instead.